// types/products-type.ts
export type ProductType = {
  productId: string;
  name: string;
  description: string;
  price: string;
  images: string;
  category: string; // Danh mục sản phẩm
  brand: string; // Thương hiệu
  size: string; // Kích thước
  color: string; // Màu sắc
  material: string; // Chất liệu
  rating: string; // Đánh giá
  reviews: string; // Đánh giá từ người dùng
  discount: string; // Giảm giá
  stock: string; // Tồn kho
  saleCount: string;
};